+++ packages/dart_mappable/test/external_types/unknown_interface.mapper.dart
@@ -43,7 +43,7 @@ class HOpacityMapper extends EnumMapper<HOpacity> {
 extension HOpacityMapperExtension on HOpacity {
   String toValue() {
     HOpacityMapper.ensureInitialized();
-    return MapperContainer.globals.toValue(this) as String;
+    return MapperContainer.globals.toValue<HOpacity>(this) as String;
   }
 }