+++ packages/dart_mappable_builder/lib/src/generators/class_mapper_generator.dart
@@ -114,11 +114,9 @@ class ClassMapperGenerator extends MapperGenerator<TargetClassMapperElement>
 
     var node = await element.element.getNode();
     if (node is ClassDeclaration) {
-      var hasCopyWithMixin = node.withClause?.mixinTypes.any((t) {
-            // ignore: deprecated_member_use
-            return t.name.name == mixinName;
-          }) ??
-          false;
+      var hasCopyWithMixin =
+          node.withClause?.mixinTypes.any((t) => t.name2.lexeme == mixinName) ??
+              false;
 
       if (!hasCopyWithMixin) {
         var classDeclarationSource = 'class $className';
@@ -140,10 +138,8 @@ class ClassMapperGenerator extends MapperGenerator<TargetClassMapperElement>
         warnUnusedMixin(classDeclarationSource);
       }
     } else if (node is ClassTypeAlias) {
-      var hasCopyWithMixin = node.withClause.mixinTypes.any((t) {
-        // ignore: deprecated_member_use
-        return t.name.name == mixinName;
-      });
+      var hasCopyWithMixin =
+          node.withClause.mixinTypes.any((t) => t.name2.lexeme == mixinName);
 
       if (!hasCopyWithMixin) {
         var classDeclarationSource = 'class $className';