+++ packages/dart_mappable/CHANGELOG.md
@@ -17,6 +17,10 @@
     
   For a more detailed usage see the documentation.
 
+# 3.3.0
+
+- Make `DateTime` encoding configurable through `DateTimeMapper.encodingMode`.
+
 # 3.2.0
 
 - Handle bounded nested type parameters in subclasses.