+++ packages/dart_mappable_builder/CHANGELOG.md
@@ -17,6 +17,11 @@
     
   For a more detailed usage see the documentation.
 
+# 3.3.0
+
+- Make enum encoding explicitly typed.
+- Update analyzer to '>=5.11.0 <7.0.0'
+
 # 3.2.3
 
 - Respect keys and hooks for annotated getters.